Define CommitmentReceived type in communication protocol
